# specs/004-pricing-tier-config/contracts/api.yaml
openapi: 3.0.3
info:
  title: membran-app Pricing Tier API
  description: API for managing pricing tier configurations for Discord server premium memberships
  version: 1.0.0
  contact:
    name: membran-app
    url: https://membran.app
servers:
  - url: http://localhost:8787
    description: Local development server
  - url: https://api.membran.app
    description: Production server
...
paths:
...
  /api/roles/sync:
    post:
      summary: Sync Discord roles from the connected server
      operationId: syncDiscordRoles
      tags: [roles]
      security:
        - CookieAuth: []
      responses:
        '200':
          description: Roles synced successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  synced:
                    type: boolean
                  roles:
                    type: array
                    items:
                      $ref: '#/components/schemas/DiscordRole'
                  warnings:
                    type: array
                    items:
                      type: string
                  message:
                    type: string
              example:
                synced: true
                roles: []
                warnings: []
                message: "Synced 15 roles from Discord"
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
        '503':
          description: Discord API unavailable - roles may be stale
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              example:
                error: DISCORD_API_UNAVAILABLE
                message: Discord is currently unavailable. Using cached role data.
        '500':
          $ref: '#/components/responses/InternalServerError'
  /api/roles:
    get:
      summary: Get cached Discord roles for the user's server
      operationId: listDiscordRoles
      tags: [roles]
      security:
        - CookieAuth: []
      responses:
        '200':
          description: List of Discord roles
          content:
            application/json:
              schema:
                type: object
                properties:
                  roles:
                    type: array
                    items:
                      $ref: '#/components/schemas/DiscordRole'
                  lastSynced:
                    type: string
                    format: date-time
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalServerError'
components:
  securitySchemes:
    CookieAuth:
      type: apiKey
      in: cookie
      name: auth_session
      description: Session cookie for authentication
...
  schemas:
...
    DiscordRole:
      type: object
      properties:
        id:
          type: string
          example: role_abc123
        discordServerId:
          type: string
          example: server_def456
        discordRoleId:
          type: string
          example: "987654321098765432"
        roleName:
          type: string
          example: Premium Member
        botCanManage:
          type: boolean
          example: true
          description: Whether bot has permission to assign this role
        position:
          type: integer
          example: 10
        color:
          type: integer
          nullable: true
          example: 16777215
        hoist:
          type: boolean
          example: false
        permissions:
          type: string
          nullable: true
          example: "8,1024,2048"
        syncedAt:
          type: string
          format: date-time
      required:
        - id
        - discordServerId
        - discordRoleId
        - roleName
        - botCanManage
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
          description: Error code for programmatic handling
          enum:
            - TIER_LIMIT_EXCEEDED
            - FEATURE_LIMIT_EXCEEDED
            - INVALID_PRICE_RANGE
            - DUPLICATE_TIER_NAME
            - ROLE_CANNOT_BE_MANAGED
            - TIER_HAS_ACTIVE_SUBSCRIBERS
            - LAST_TIER_CANNOT_DELETE
            - VERSION_CONFLICT
            - VALIDATION_ERROR
            - UNAUTHORIZED
            - NOT_FOUND
            - INTERNAL_ERROR
        message:
          type: string
          description: Human-readable error message
        details:
          type: object
          description: Additional error details
          additionalProperties: true
      required:
        - error
        - message
  responses:
...
    Unauthorized:
      description: Unauthorized - invalid or missing session
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            error: UNAUTHORIZED
            message: No valid session
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            error: NOT_FOUND
            message: Pricing tier not found
...
    InternalServerError:
      description: Internal server error
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            error: INTERNAL_ERROR
            message: An unexpected error occurred